Only update selection to marker on form open
