+++ app/assets/javascripts/components/annotations/code_listing_row.ts
@@ -1,98 +1,22 @@
 import { ShadowlessLitElement } from "components/meta/shadowless_lit_element";
 import { customElement, property } from "lit/decorators.js";
-import { html, TemplateResult, render } from "lit";
+import { html, TemplateResult } from "lit";
 import { unsafeHTML } from "lit/directives/unsafe-html.js";
 import "components/annotations/hidden_annotations_dot";
 import "components/annotations/annotations_cell";
 import "components/annotations/annotation_marker";
 import { i18nMixin } from "components/meta/i18n_mixin";
-import { initTooltips, sleep } from "util.js";
+import { initTooltips } from "util.js";
 import { PropertyValues } from "@lit/reactive-element";
 import { userState } from "state/Users";
-import { AnnotationData, annotationState } from "state/Annotations";
+import { AnnotationData } from "state/Annotations";
 import { MachineAnnotationData, machineAnnotationState } from "state/MachineAnnotations";
 import { wrapRangesInHtml, range } from "mark";
 import { SelectedRange, UserAnnotationData, userAnnotationState } from "state/UserAnnotations";
 import { AnnotationMarker } from "components/annotations/annotation_marker";
 import "components/annotations/selection_marker";
 import "components/annotations/selection_tooltip";
-
-function getOffset(e: Node, o: number): number | undefined {
-    if (e.nodeName === "PRE") {
-        return o;
-    }
-
-    const parent = e.parentNode;
-    if (!parent) {
-        return undefined;
-    }
-
-    let precedingText = "";
-    for (const child of parent.childNodes) {
-        if (child === e) {
-            break;
-        }
-        if (child.nodeType !== Node.COMMENT_NODE) {
-            precedingText += child.textContent;
-        }
-    }
-    return getOffset(parent, o + precedingText.length);
-}
-
-function selectedRangeFromSelection(s: Selection): SelectedRange | undefined {
-    // Selection.anchorNode does not behave as expected in firefox, see https://bugzilla.mozilla.org/show_bug.cgi?id=1420854
-    // So we use the startContainer of the range instead
-    const anchorNode = s.getRangeAt(0).startContainer;
-    const focusNode = s.getRangeAt(s.rangeCount - 1).endContainer;
-    const anchorOffset = s.getRangeAt(0).startOffset;
-    const focusOffset = s.getRangeAt(s.rangeCount - 1).endOffset;
-
-    const anchorRow = anchorNode?.parentElement.closest("d-code-listing-row") as CodeListingRow;
-    const focusRow = focusNode?.parentElement.closest("d-code-listing-row") as CodeListingRow;
-    const anchorColumn = getOffset(anchorNode, anchorOffset);
-    const focusColumn = getOffset(focusNode, focusOffset);
-
-    if (!anchorRow || !focusRow || anchorColumn === undefined || focusColumn === undefined) {
-        return undefined;
-    }
-
-    let range: SelectedRange;
-    if (anchorRow.row < focusRow.row) {
-        range = {
-            row: anchorRow.row,
-            rows: focusRow.row - anchorRow.row + 1,
-            column: anchorColumn,
-            columns: focusColumn,
-        };
-    } else if (anchorRow.row > focusRow.row) {
-        range = {
-            row: focusRow.row,
-            rows: anchorRow.row - focusRow.row + 1,
-            column: focusColumn,
-            columns: anchorColumn,
-        };
-    } else if (anchorColumn < focusColumn) {
-        range = {
-            row: anchorRow.row,
-            rows: 1,
-            column: anchorColumn,
-            columns: focusColumn - anchorColumn,
-        };
-    } else {
-        range = {
-            row: anchorRow.row,
-            rows: 1,
-            column: focusColumn,
-            columns: anchorColumn - focusColumn,
-        };
-    }
-
-    if (range.columns === 0 && range.rows > 1) {
-        range.columns = undefined;
-        range.rows -= 1;
-    }
-    return range;
-}
+import { triggerSelectionStart } from "components/annotations/select";
 
 /**
  * This component represents a row in the code listing.
@@ -111,32 +35,6 @@ export class CodeListingRow extends i18nMixin(ShadowlessLitElement) {
     @property({ type: String })
     renderedCode: string;
 
-    async triggerSelectionEnd(): Promise<void> {
-        console.log("triggerSelectionEnd");
-        document.body.classList.remove("no-selection-outside-code");
-        if (userAnnotationState.showForm) {
-            return;
-        }
-
-        // Wait for the selection to be updated
-        await sleep(10);
-        const selection = window.getSelection();
-        if (!selection.isCollapsed) {
-            userAnnotationState.selectedRange = selectedRangeFromSelection(selection);
-            if (userAnnotationState.selectedRange) {
-                selection.removeAllRanges();
-            }
-        } else {
-            userAnnotationState.selectedRange = undefined;
-        }
-    }
-
-    triggerSelectionStart(e: PointerEvent): void {
-        if (!(e.target as Element).closest(".annotation")) {
-            console.log("triggerSelectionStart");
-            document.body.classList.add("no-selection-outside-code");
-        }
-    }
 
     /**
      * Calculates the range of the code that is covered by the given annotation.
@@ -183,7 +81,7 @@ export class CodeListingRow extends i18nMixin(ShadowlessLitElement) {
                 annotations.push(range.data);
                 node.setAttribute("annotations", JSON.stringify(annotations));
             });
-        if (userAnnotationState.selectedRange && userAnnotationState.selectedRange.row <= this.row && userAnnotationState.selectedRange.row + (userAnnotationState.selectedRange.rows ?? 1) > this.row) {
+        if ( userAnnotationState.showForm && userAnnotationState.selectedRange && userAnnotationState.selectedRange.row <= this.row && userAnnotationState.selectedRange.row + (userAnnotationState.selectedRange.rows ?? 1) > this.row) {
             annotationsMarked = wrapRangesInHtml(annotationsMarked, [this.getRangeFromAnnotation(userAnnotationState.selectedRange)], "d-selection-marker");
         }
         return annotationsMarked;
@@ -192,8 +90,7 @@ export class CodeListingRow extends i18nMixin(ShadowlessLitElement) {
     firstUpdated(_changedProperties: PropertyValues): void {
         super.firstUpdated(_changedProperties);
         initTooltips(this);
-        this.addEventListener("pointerup", () => this.triggerSelectionEnd());
-        this.addEventListener("pointerdown", e => this.triggerSelectionStart(e));
+        this.addEventListener("pointerdown", e => triggerSelectionStart(e));
     }
 
     get canCreateAnnotation(): boolean {
